requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2==2.9.6
pycodestyle==2.10.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
